Adjusted variables type

+++ LinearEq.java
@@ -1,8 +1,8 @@
 class LinearEq {
 	public static void main (String [] args){
-		double a = Integer.parseInt(args[0]);
-		double b = Integer.parseInt(args[1]);
-		double c = Integer.parseInt(args[2]);	
+		double a = Double.parseDouble(args[0]);
+		double b = Double.parseDouble(args[1]);
+		double c = Double.parseDouble(args[2]);	
 
 		double SolveForX = (c-b)/a;
